fix responsive issues with topics and media context
